refactoring: create lib dir
