+++ src/index.ts
@@ -1,178 +1,25 @@
-import { FIREBASE_FUNCTIONS_URL, GARBAGE_FRIDAY, GARBAGE_MONDAY, GARBAGE_SATURDAY, GARBAGE_SUNDAY, GARBAGE_THURSDAY, GARBAGE_TUESDAY, GARBAGE_WEDNESDAY, LINE_CHANNEL_ACCESS_TOKEN_HOME, LINE_CHANNEL_ACCESS_TOKEN_NPB, LINE_GROUP_ID, LINE_USER_ID, SLACK_USER_ID, SLACK_WEBHOOK_URL } from "./constants";
-import { GameInfo, ResultType, SlackBlockInfo, SlackPayloadType } from "./types";
-
-
-/**
- * 試合情報の取得
- * @returns {ResultType}
- */
-function getGameInfo(): ResultType {
-  const rawGameInfo: string = UrlFetchApp.fetch(FIREBASE_FUNCTIONS_URL).getContentText();
-  const gameInfo: ResultType = JSON.parse(rawGameInfo);
-  return gameInfo;
-}
+import { getGameInfo, organizeGameInfo, postToSlack, organizeGameInfoForLine, getTomorrowGarbageInfo, getGarbageInfoList } from "./lib";
+import { LINE_CHANNEL_ACCESS_TOKEN_NPB, LINE_CHANNEL_ACCESS_TOKEN_HOME, LINE_GROUP_ID, LINE_USER_ID } from "./constants";
 
 /**
- * 試合情報の整理
- * @param {ResultType} fetchedResult
- * @returns {SlackPayloadType}
+ * Slack へ NPB の試合一覧を通知する関数
  */
-function organizeGameInfo(fetchedResult: ResultType): SlackPayloadType {
-  const {
-    dateInfo,
-    gameInfoList,
-  } = fetchedResult;
-  const payload: SlackPayloadType = {
-    blocks: []
+function notifyToSlackAboutNPB() {
+  try {
+    const gameInfo = getGameInfo();
+    console.log(gameInfo);
+    const organizedGameInfo = organizeGameInfo(gameInfo);
+    console.log(organizedGameInfo);
+    postToSlack(organizedGameInfo);
+  } catch (error) {
+    console.log(error);
   }
-  const mentionBlockInfo: SlackBlockInfo = {
-    type: "section",
-    text: {
-      type: "mrkdwn",
-      text: `<@${SLACK_USER_ID}>\n\n`
-    }
-  };
-  payload.blocks.push(mentionBlockInfo);
-
-  if (dateInfo && gameInfoList && gameInfoList.length > 0) {
-    const dateBlockInfo: SlackBlockInfo = {
-      type: "section",
-      text: {
-        type: "mrkdwn",
-        text: `${dateInfo}\n\n`
-      }
-    }
-    payload.blocks.push(dateBlockInfo);
-
-    for (let i = 0; i < gameInfoList.length; i ++) {
-      const gameInfo: GameInfo = gameInfoList[i];
-      const {
-        leftTeamName,
-        rightTeamName,
-        leftTeamScore,
-        rightTeamScore,
-        gameStateInfo
-      } = gameInfo;
-      if (leftTeamName && rightTeamName) {
-        if (leftTeamScore && rightTeamScore) {
-          const gameInfo = `${leftTeamName} ${leftTeamScore} - ${rightTeamScore} ${rightTeamName}`;
-          if (gameStateInfo) {
-            const gameBlockInfo: SlackBlockInfo = {
-              type: "section",
-              text: {
-                type: "mrkdwn",
-                text: `• ${gameInfo + "\n" + gameStateInfo + "\n"}`
-              }
-            };
-            payload.blocks.push(gameBlockInfo);
-          } else {
-            const gameBlockInfo: SlackBlockInfo = {
-              type: "section",
-              text: {
-                type: "mrkdwn",
-                text: `• ${gameInfo + "\n"}`
-              }
-            };
-            payload.blocks.push(gameBlockInfo);
-          }
-        } else {
-          const gameInfo = `${leftTeamName} vs ${rightTeamName}`;
-          if (gameStateInfo) {
-            const gameBlockInfo: SlackBlockInfo = {
-              type: "section",
-              text: {
-                type: "mrkdwn",
-                text: `• ${gameInfo + "\n" + gameStateInfo + "\n"}`
-              }
-            };
-            payload.blocks.push(gameBlockInfo);
-          } else {
-            const gameBlockInfo: SlackBlockInfo = {
-              type: "section",
-              text: {
-                type: "mrkdwn",
-                text: `• ${gameInfo + "\n"}`
-              }
-            };
-            payload.blocks.push(gameBlockInfo);
-          }
-        }
-      }
-    }
-  } else {
-    const noGameDayInfo: SlackBlockInfo = {
-      type: 'section',
-      text: {
-        type: 'mrkdwn',
-        text: '今日は試合がありません'
-      }
-    };
-    payload.blocks.push(noGameDayInfo);
-  }
-  return payload;
-}
-
-/**
- * Slack への通知
- * @param {SlackPayloadType} organizedGameInfo
- */
-function postToSlack(organizedGameInfo: SlackPayloadType) {
-  const options = {
-    'method' : 'post',
-    'contentType': 'application/json',
-    'payload' : JSON.stringify(organizedGameInfo)
-  };
-  const response = UrlFetchApp.fetch(SLACK_WEBHOOK_URL, options);
-  console.log(response);
-}
-
-/**
- * LINE 用のメッセージ整形
- * @param {ResultType} fetchedResult
- * @returns {string}
- */
-function organizeGameInfoForLine(fetchedResult: ResultType): string {
-  const {
-    dateInfo,
-    gameInfoList,
-  } = fetchedResult;
-  let organizedGameInfoForLine = '';
-  organizedGameInfoForLine += `${dateInfo}\n\n`;
-
-  for (let i = 0; i < gameInfoList.length; i ++) {
-    const gameInfo: GameInfo = gameInfoList[i];
-    const {
-      leftTeamName,
-      rightTeamName,
-      leftTeamScore,
-      rightTeamScore,
-      gameStateInfo
-    } = gameInfo;
-    if (leftTeamName && rightTeamName) {
-      if (leftTeamScore && rightTeamScore) {
-        const gameInfo = `${leftTeamName} ${leftTeamScore} - ${rightTeamScore} ${rightTeamName}`;
-        if (gameStateInfo) {
-          organizedGameInfoForLine += `• ${gameInfo + "\n" + gameStateInfo + "\n\n"}`
-        } else {
-          organizedGameInfoForLine += `• ${gameInfo + "\n\n"}`
-        }
-      } else {
-        const gameInfo = `${leftTeamName} vs ${rightTeamName}`;
-        if (gameStateInfo) {
-          organizedGameInfoForLine += `• ${gameInfo + "\n" + gameStateInfo + "\n\n"}`
-        } else {
-          organizedGameInfoForLine += `• ${gameInfo + "\n\n"}`
-        }
-      }
-    }
-  }
-  return organizedGameInfoForLine;
 }
 
 /**
  * LINE への通知
  */
-function postToLINE() {
+function notifyToLINEAboutNPB() {
   try {
     const gameInfo = getGameInfo();
     console.log(gameInfo);
@@ -204,67 +51,10 @@ function postToLINE() {
   }
 }
 
-function getTomorrowGarbageInfo() {
-  const todayWeekNum = new Date().getDay();
-  const tomorrowWeekNum = todayWeekNum === 6
-    ? 0
-    : todayWeekNum + 1;
-  const weekDays = [
-    "日曜日",
-    "月曜日",
-    "火曜日",
-    "水曜日",
-    "木曜日",
-    "金曜日",
-    "土曜日"
-  ];
-  const garbages = [
-    GARBAGE_SUNDAY,
-    GARBAGE_MONDAY,
-    GARBAGE_TUESDAY,
-    GARBAGE_WEDNESDAY,
-    GARBAGE_THURSDAY,
-    GARBAGE_FRIDAY,
-    GARBAGE_SATURDAY,
-  ]
-  const tomorrowDay = weekDays[tomorrowWeekNum];
-  const tomorrowGarbage = garbages[tomorrowWeekNum];
-  const message = `明日は${tomorrowDay}
-${tomorrowGarbage} の日
-
-  `
-  console.log(message);
-  return message;
-}
-
-/**
- * ゴミ出しの情報一覧取得
- */
-function getGarbageInfoList() {
-  const MondayInfo = GARBAGE_MONDAY;
-  const TuesdayInfo = GARBAGE_TUESDAY;
-  const WednesdayInfo = GARBAGE_WEDNESDAY;
-  const ThursdayInfo = GARBAGE_THURSDAY;
-  const FridayInfo = GARBAGE_FRIDAY;
-  const SaturdayInfo = GARBAGE_SATURDAY;
-  const SundayInfo = GARBAGE_SUNDAY;
-  const message = `ゴミ出しの日一覧
-
-  月曜日: ${MondayInfo}
-  火曜日: ${TuesdayInfo}
-  水曜日: ${WednesdayInfo}
-  木曜日: ${ThursdayInfo}
-  金曜日: ${FridayInfo}
-  土曜日: ${SaturdayInfo}
-  日曜日: ${SundayInfo}
-  `
-  return message;
-}
-
 /**
  * ゴミ出しの通知
  */
-function notifyAboutGarbage() {
+ function notifyToLINEAboutGarbage() {
   try {
     const url = 'https://api.line.me/v2/bot/message/push';
     const tomorrowInfo = getTomorrowGarbageInfo();
@@ -290,17 +80,3 @@ function notifyAboutGarbage() {
   }
 }
 
-/**
- * メイン関数
- */
-function main() {
-  try {
-    const gameInfo = getGameInfo();
-    console.log(gameInfo);
-    const organizedGameInfo = organizeGameInfo(gameInfo);
-    console.log(organizedGameInfo);
-    postToSlack(organizedGameInfo);
-  } catch (error) {
-    console.log(error);
-  }
-}
